+++ webapp-frontend/src/app/services/ui/notification.service.ts
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -32,20 +32,20 @@ export class NotificationService {
     duration: 3000,
     horizontalPosition: 'right',
     verticalPosition: 'top'
-  }
+  };
 
 
-  success(msg) {
+  success(msg: string) {
     this.config['panelClass'] = ['notification', 'success', 'default'];
     this.snackBar.open(msg, '', this.config);
   }
 
-  warn(msg) {
+  warn(msg: string) {
     this.config['panelClass'] = ['notification', 'warn', 'default'];
     this.snackBar.open(msg, '', this.config);
   }
 
-  error(msg) {
+  error(msg: string) {
     this.config['panelClass'] = ['notification', 'error', 'default'];
     this.snackBar.open(msg, '', this.config);
   }