+++ docs/release-notes.rst
@@ -27,6 +27,8 @@ Version 1.0.3, 20 May 2019
 * Update ANR xApp client to spec version 0.0.5
 * Add get-version methods to all controllers
 * Add simple page footer with copyright and version
+* Add AC and ANR xApp services
+* Rename signal service to E2 Manager service
 
 Version 1.0.2, 13 May 2019
 --------------------------